issue: 1928915 Keep pcb->last_unsent valid
When tcp_output() sends the last segment it leaves pcb->last_unsent
outdated. It can happen that last_unsent points to a freed segment
which is the last segment in the free list. In turn, it means that
last_unsent->next is NULL.

TCP code doesn't rely on validity of pcb->last_unsent pointer and
updates it every time it doesn't meet certain conditions. And the
pointer is NOT updated when pcb->last_unsent->next is NULL.
Therefore, in the above scenario, a freed segment can be updated
as the last unsent segment what corrupts the free list.

To resolve the issue, keep pcb->last_unsent pointer valid. It also
avoids segfault or wrong result in tcp_is_wnd_available().

As further step, the redundant code which updates pcb->last_unsent
can be removed.

Signed-off-by: Dmytro Podgornyi <dmytrop@mellanox.com>